requirements.txt
django
djangorestframework
django-xadmin
